#89 Move helpers to the view
